remove box model
The whole box model thing gets really complex, really fast.
For an initial release, it makes more sense to release something
simple, that works and is easily explained. If someone feels
ambitious, they are welcome to work on a more advanced model
that actually supports more of HTML and/or CSS.